split structs
